refactor(core,gateway): split Stores.swift by domain and Routing/ by subsystem
Stores.swift packed ~15 unrelated store protocols into one 486-line file;
each domain family now gets its own file under ClawCore/Persistence/, with
the result structs that belong to it. ClawGateway/Routing/ mixed four
subsystems across 19 files; the approval fabric, schedule surface, and turn
execution move to Approval/, Schedule/, and Turn/. Moves only — no
declaration changes.
