+++ Sources/ClawCore/Persistence/AuditLog.swift
@@ -1,0 +1,3 @@
+public protocol AuditLog: Sendable {
+  func appendAudit(_ event: AuditEvent) throws
+}
