+++ Sources/ClawCore/Persistence/IntakeStores.swift
@@ -1,0 +1,16 @@
+public protocol AllowlistStore: Sendable {
+  func seedAllowlist(userIds: [Int64]) throws
+  func allowlistContains(userId: Int64) throws -> Bool
+  func allowlistCount() throws -> Int
+}
+
+public protocol ProcessedUpdateStore: Sendable {
+  /// INSERT OR IGNORE → true if newly claimed, false if already seen.
+  /// Synchronous: no await may span the check, so the dedup claim can't interleave.
+  func claimUpdate(updateId: Int64) throws -> Bool
+}
+
+public protocol UpdateCursorStore: Sendable {
+  func loadCursor() throws -> Int64?
+  func advanceCursor(to updateId: Int64) throws
+}
