+++ Sources/ClawGateway/Approval/ApprovalCoordinator.swift
