+++ Sources/ClawGateway/Approval/ApprovedActionExecutor.swift
