+++ Sources/ClawGateway/Approval/ConfirmationResolver.swift
