+++ Sources/ClawGateway/Approval/PendingConfirmationRegistry.swift
