+++ Sources/ClawGateway/Schedule/ScheduleDraftParser.swift
