+++ Sources/ClawGateway/Turn/BudgetBreaker.swift
