+++ Sources/ClawGateway/Turn/TurnRunner+Payloads.swift
